Linking is hard.
After a few hours of struggling with linking, it turns out my Windows build of Clang was half of the problem.

Not quite as painful as I had imagined, but I got a make-system setup which allows seamlessly building the c-unittest.

Now that all of that is settled, it should be trivial to set up the rest of the c-unittest.